patch Rails::DBConsole
